refactor: make separate hook for storing and mutating the data
